I init convertBigIntToString lib.
