+++ lib/stack/trackEventWithStack.js
@@ -2,36 +2,12 @@ import fetch from "node-fetch";
 import dotenv from "dotenv";
 import { decodeEventLog } from "viem";
 import { zoraCreator1155ImplABI } from "@zoralabs/protocol-deployments";
+import convertBigIntToString from "../convertBigIntToString.js";
 
 dotenv.config();
 
 export const STACK_API_KEY = process.env.STACK_API_KEY;
 export const STACK_API_URL = "https://track.stack.so/event";
-
-// Helper function to convert BigInt to string
-function convertBigIntToString(obj) {
-  if (typeof obj !== "object" || obj === null) {
-    if (typeof obj === "string" && obj.startsWith("0x")) {
-      // Convert hexadecimal strings to decimal strings
-      return BigInt(obj).toString(10);
-    }
-    if (typeof obj === "bigint") {
-      return obj.toString(10);
-    }
-    return obj;
-  }
-
-  if (Array.isArray(obj)) {
-    return obj.map(convertBigIntToString);
-  }
-
-  return Object.fromEntries(
-    Object.entries(obj).map(([key, value]) => [
-      key,
-      convertBigIntToString(value),
-    ])
-  );
-}
 
 async function trackEventWithStack(network, event) {
   try {
